requirements.txt
python-binance==1.0.29
python-dotenv>=1.0
pandas>=2.2
numpy>=1.26